Style sheets of Accounts, ForgotPass, Login, Register, [AddonScreen, Help, HotelPage, ReviewPage are updated with bugs fixed......Order Page has problems with layout is now fixed, there is a little issue that permission for camera sometimes needs to be reloaded, I made some changes, could be fixed ]:Moeez Ahmed
