+++ Styles/AddOnStyles.js
@@ -1,0 +1,63 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#f8f8f8",
+    padding: 16,
+  },
+  header: {
+    fontSize: 24,
+    fontWeight: "bold",
+    color: "#333",
+    marginBottom: 20,
+    textAlign: "center",
+  },
+  addonItem: {
+    backgroundColor: "#fff",
+    padding: 12,
+    borderRadius: 8,
+    marginVertical: 8,
+    borderWidth: 1,
+    borderColor: "#ddd",
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  addonItemSelected: {
+    backgroundColor: "#f0f8ff",
+    borderColor: "#4caf50",
+    borderWidth: 2,
+  },
+  addonText: {
+    fontSize: 16,
+    color: "#333",
+  },
+  addonPrice: {
+    fontSize: 16,
+    color: "#666",
+  },
+  doneButton: {
+    backgroundColor: "#4caf50",
+    paddingVertical: 12,
+    borderRadius: 8,
+    marginTop: 20,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  doneButtonText: {
+    fontSize: 18,
+    color: "#fff",
+    fontWeight: "bold",
+  },
+  loadingText: {
+    fontSize: 18,
+    color: "#888",
+    textAlign: "center",
+  },
+  emptyText: {
+    textAlign: "center",
+    fontSize: 16,
+    color: "#888",
+  },
+});
