+++ Styles/ForgotPassStyles.js
@@ -1,0 +1,72 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  safeAreaContainer: {
+    flex: 1,
+    backgroundColor: 'white',
+  },
+  backButton: {
+    marginTop: 20,
+    marginLeft: 20,
+    width: 40,
+    height: 40,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  mainContainer: {
+    flex: 1,
+    paddingHorizontal: 20,
+    justifyContent: 'center',
+  },
+  headerText: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    marginBottom: 15,
+    color: '#333',
+    textAlign: 'center',
+  },
+  subHeaderText: {
+    fontSize: 16,
+    color: 'gray',
+    marginBottom: 30,
+    textAlign: 'center',
+    lineHeight: 24,
+  },
+  inputContainer: {
+    marginBottom: 20,
+  },
+  label: {
+    fontSize: 14,
+    color: 'gray',
+    marginBottom: 5,
+  },
+  emailViewContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: '#e0e0e0',
+    borderRadius: 5,
+    paddingHorizontal: 10,
+    paddingVertical: 5,
+  },
+  icon: {
+    marginRight: 10,
+  },
+  emailInput: {
+    flex: 1,
+    color: 'gray',
+  },
+  resetButton: {
+    backgroundColor: '#fd5c63',
+    borderRadius: 6,
+    paddingVertical: 15,
+    alignItems: 'center',
+    marginTop: 20,
+  },
+  resetButtonText: {
+    color: 'white',
+    fontSize: 16,
+    fontWeight: 'bold',
+  },
+})
+
+export default styles
