+++ Styles/HelpStyles.js
@@ -1,0 +1,94 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    paddingTop: 20,
+    backgroundColor: '#f7f7f7',
+  },
+  header: {
+    backgroundColor: '#fff',
+    padding: 15,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  headerText: {
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+  backButton: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  backText: {
+    marginLeft: 10,
+  },
+  scrollView: {
+    flex: 1,
+  },
+  contentContainer: {
+    padding: 15,
+  },
+  centeredText: {
+    textAlign: 'center',
+  },
+  subHeaderText: {
+    fontSize: 16,
+    color: '#555',
+  },
+  sectionContainer: {
+    backgroundColor: 'white',
+    borderRadius: 10,
+    marginBottom: 10,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  sectionHeader: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    padding: 15,
+  },
+  expandButton: {
+    color: '#fc8019',
+  },
+  sectionContent: {
+    padding: 15,
+  },
+  tipTitle: {
+    fontSize: 18,
+    fontWeight: 'bold',
+  },
+  tipDescription: {
+    fontSize: 14,
+    color: '#777',
+  },
+  contactContainer: {
+    backgroundColor: 'white',
+    borderRadius: 10,
+    padding: 15,
+    alignItems: 'center',
+    marginTop: 15,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  contactButton: {
+    backgroundColor: '#fc8019',
+    paddingVertical: 10,
+    paddingHorizontal: 20,
+    borderRadius: 8,
+  },
+  contactButtonText: {
+    color: 'white',
+    fontWeight: 'bold',
+  },
+});
