+++ Styles/orderStyles.js
@@ -1,152 +1,180 @@
-import { StyleSheet } from "react-native";
+import { StyleSheet } from 'react-native';
 
 export const styles = StyleSheet.create({
     container: {
         flex: 1,
-        backgroundColor: "white",
+        backgroundColor: '#fff',
+        padding: 10,
     },
     header: {
-        flexDirection: "row",
-        alignItems: "center",
-        justifyContent: "space-between",
-        height: 60,
-        backgroundColor: "#fd5c63",
-        paddingHorizontal: 15,
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        marginBottom: 20,
     },
     headerText: {
-        color: "white",
-        fontSize: 16,
-        fontWeight: "bold",
+        fontSize: 18,
+        fontWeight: 'bold',
     },
     subHeaderText: {
-        color: "white",
         fontSize: 14,
+        color: '#888',
     },
     helpText: {
-        color: "white",
-        fontSize: 15,
-        fontWeight: "bold",
-    },
-    mapContainer: {
-        position: "relative",
-    },
-    map: {
-        width: "100%",
-        height: 400,
-    },
-    centerButton: {
-        position: "absolute",
-        bottom: 15,
-        right: 15,
-        backgroundColor: "#fd5c63",
-        borderRadius: 25,
-        flexDirection: "row",
-        alignItems: "center",
-        justifyContent: "center",
-        paddingHorizontal: 15,
-        paddingVertical: 10,
-        elevation: 5,
-    },
-    centerButtonText: {
-        color: "white",
-        marginLeft: 5,
-        fontWeight: "bold",
+        color: '#fc8019',
+        fontWeight: 'bold',
         fontSize: 14,
     },
-    tipContainer: {
-        height: 320,
-        backgroundColor: "white",
-        borderTopLeftRadius: 20,
-        borderTopRightRadius: 20,
-        padding: 15,
-        shadowColor: "#000",
-        shadowOffset: { width: 0, height: -2 },
+    mapContainer: {
+        height: '50%',
+        marginBottom: 20,
+    },
+    map: {
+        width: '100%',
+        height: '100%',
+    },
+    centerButton: {
+        position: 'absolute',
+        bottom: 20,
+        left: '50%',
+        marginLeft: -60,
+        backgroundColor: '#4A89F3',
+        padding: 10,
+        borderRadius: 50,
+        alignItems: 'center',
+        justifyContent: 'center',
+        flexDirection: 'row',
+        shadowColor: '#000',
+        shadowOffset: { width: 0, height: 2 },
         shadowOpacity: 0.2,
-        shadowRadius: 5,
-        elevation: 10,
+        shadowRadius: 3,
+        elevation: 3,
+    },
+    centerButtonText: {
+        color: 'white',
+        fontSize: 12,
+        marginLeft: 5,
+    },
+    errorContainer: {
+        position: 'absolute',
+        bottom: 10,
+        left: '50%',
+        marginLeft: -150,
+        backgroundColor: '#ffdddd',
+        padding: 10,
+        borderRadius: 5,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    errorText: {
+        color: '#ff0000',
+        fontWeight: 'bold',
+    },
+    tipContainer: {
+        flex: 1,
+        justifyContent: 'flex-end',
+        paddingBottom: 20,
     },
     tipContent: {
-        flex: 1,
+        backgroundColor: '#f9f9f9',
+        padding: 15,
+        borderRadius: 10,
+        shadowColor: '#000',
+        shadowOffset: { width: 0, height: 2 },
+        shadowOpacity: 0.1,
+        shadowRadius: 3,
+        elevation: 3,
     },
     tipTextCenter: {
-        fontSize: 18,
-        fontWeight: "bold",
-        textAlign: "center",
-        marginBottom: 15,
+        textAlign: 'center',
+        fontSize: 16,
+        fontWeight: 'bold',
+        marginBottom: 10,
     },
     tipRow: {
-        flexDirection: "row",
-        alignItems: "center",
-        marginVertical: 15,
+        flexDirection: 'row',
+        alignItems: 'center',
+        marginTop: 15,
     },
     tipDetails: {
-        marginLeft: 15,
+        marginLeft: 10,
         flex: 1,
     },
     tipTitle: {
-        fontSize: 16,
-        fontWeight: "bold",
-        marginBottom: 5,
+        fontSize: 14,
+        fontWeight: 'bold',
     },
     tipDescription: {
-        fontSize: 14,
-        color: "#7d7d7d",
+        fontSize: 12,
+        color: '#555',
+        marginVertical: 5,
     },
     tipOptions: {
-        flexDirection: "row",
-        marginTop: 15,
+        flexDirection: 'row',
+        justifyContent: 'space-between',
     },
     tipButton: {
-        backgroundColor: "#f5f5f5",
-        borderRadius: 10,
         padding: 10,
-        marginHorizontal: 5,
-        alignItems: "center",
-        borderWidth: 1,
-        borderColor: "#ddd",
+        borderRadius: 5,
+        backgroundColor: '#f2f2f2',
+        marginVertical: 5,
+        alignItems: 'center',
+        flexDirection: 'row',
+    },
+    selectedTipButton: {
+        backgroundColor: '#fc8019',
     },
     tipAmount: {
         fontSize: 14,
-        fontWeight: "bold",
-        color: "#002D62",
-    },
-    mostTipped: {
-        marginTop: 5,
-        fontSize: 12,
-        fontWeight: "bold",
-        color: "orange",
-        backgroundColor: "#fff0e0",
-        paddingHorizontal: 5,
-        borderRadius: 5,
-    },
-    selectedTipButton: {
-        backgroundColor: "#fd5c63",
-        borderColor: "#fd5c63",
     },
     selectedTipText: {
-        color: "white",
+        color: 'white',
+    },
+    mostTipped: {
+        fontSize: 12,
+        color: '#fc8019',
     },
     tipPayment: {
-        marginTop: 15,
         fontSize: 14,
-        color: "#fc8019",
-        textAlign: "center",
+        color: '#333',
+        marginTop: 10,
     },
     cancelTip: {
         marginTop: 5,
-        alignSelf: "center",
-        width: "40%",
-        paddingVertical: 10,
-        borderRadius: 25,
-        borderWidth: 2,
-        borderColor: "red",
-        backgroundColor: "red",
-        alignItems: "center",
-        justifyContent: "center",
+        alignItems: 'center',
     },
     cancelText: {
-        color: "white",
+        fontSize: 14,
+        color: '#fc8019',
+    },
+    cancelOrderButton: {
+        backgroundColor: '#fc8019',
+        padding: 10,
+        borderRadius: 5,
+        marginTop: 15,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    cancelOrderText: {
+        color: 'white',
+        fontWeight: 'bold',
+    },
+    orderReceivedButton: {
+        backgroundColor: '#fd5c63',
+        borderRadius: 25,
+        paddingVertical: 12,
+        paddingHorizontal: 30,
+        alignItems: 'center',
+        justifyContent: 'center',
+        shadowColor: '#000',
+        shadowOffset: { width: 0, height: 2 },
+        shadowOpacity: 0.3,
+        shadowRadius: 5,
+        elevation: 5,
+    },
+    orderReceivedText: {
+        color: 'white',
         fontSize: 16,
-        fontWeight: "bold",
+        fontWeight: 'bold',
     },
 });
