+++ app/(authenticate)/Register.js
@@ -1,8 +1,9 @@
 import React, { useState } from 'react'
-import { SafeAreaView, StyleSheet, Text, View, Pressable, TextInput, Alert, TouchableOpacity, KeyboardAvoidingView, Platform, ScrollView } from 'react-native'
+import { SafeAreaView, Text, View, Pressable, TextInput, Alert, TouchableOpacity, KeyboardAvoidingView, Platform, ScrollView } from 'react-native'
 import { supabase } from '../../lib/supabase'
 import Avatar from '../../components/Avatar'
 import { useRouter } from 'expo-router'
+import styles from '../../Styles/RegisterStyles'
 
 const Register = () => {
     const [email, setEmail] = useState('')
@@ -46,7 +47,6 @@ const Register = () => {
                 <ScrollView contentContainerStyle={{ flexGrow: 1, justifyContent: 'center' }}>
                     <View style={styles.mainView}>
                         <Text style={styles.headerTextStyle}>Food App</Text>
-                        
                     </View>
 
                     <View style={{alignItems:"center"}}>       
@@ -112,67 +112,3 @@ const Register = () => {
 }
 
 export default Register
-
-const styles = StyleSheet.create({
-    headerTextStyle: {
-        fontSize: 20,
-        textAlign: 'center',
-        fontWeight: 'bold',
-    },
-    safeAreaContainer: {
-        flex: 1,
-        backgroundColor: 'white',
-        alignItems: 'center',
-    },
-    mainView: {
-        marginTop: 50,
-    },
-    subHeadingFont: {
-        fontSize: 17,
-        fontWeight: 'bold',
-        marginTop: 12,
-        color: 'red',
-        alignItems:"center"
-    },
-    textInputView: {
-        marginTop: 30,
-        alignItems: 'center',
-    },
-    inputContainer: {
-        marginTop: 15,
-        width: 300,
-    },
-    label: {
-        fontSize: 14,
-        color: 'gray',
-        marginBottom: 5,
-    },
-    textInput: {
-        color: 'gray',
-        backgroundColor: '#e0e0e0',
-        padding: 10,
-        borderRadius: 5,
-    },
-    pressableContainer: {
-        width: 200,
-        backgroundColor: '#fd5c63',
-        borderRadius: 6,
-        marginTop: 30,
-        padding: 15,
-    },
-    registerButtonText: {
-        textAlign: 'center',
-        fontWeight: 'bold',
-        fontSize: 16,
-        color: 'white',
-    },
-    loginRedirect: {
-        marginTop: 20,
-    },
-    loginText: {
-        textAlign: 'center',
-        color: 'gray',
-        fontWeight: 'bold',
-        fontSize: 16,
-    },
-})
