+++ app/(home)/Help.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Text, View, SafeAreaView, TouchableOpacity, ScrollView, Linking } from "react-native";
 import { FontAwesome5, Ionicons } from "@expo/vector-icons";
-import { styles } from "../../Styles/orderStyles";
+import { styles } from "../../Styles/HelpStyles";  // Import styles from external file
 import { useRouter } from "expo-router";
 
 const HelpPage = () => {
@@ -48,7 +48,6 @@ const HelpPage = () => {
           return Linking.openURL(emailUrl);
         } else {
           console.log("Can't handle email URL");
-          // Optionally, show an alert that email app couldn't be opened
         }
       })
       .catch((err) => console.error('An error occurred', err));
@@ -60,87 +59,55 @@ const HelpPage = () => {
       <View style={styles.header}>
         <TouchableOpacity 
           onPress={() => router.back()} 
-          style={{ flexDirection: 'row', alignItems: 'center' }}
+          style={styles.backButton}
         >
           <Ionicons name="arrow-back" size={24} color="black" />
-          <Text style={[styles.headerText, { marginLeft: 10 }]}>Back</Text>
+          <Text style={[styles.headerText, styles.backText]}>Back</Text>
         </TouchableOpacity>
       </View>
 
-      <ScrollView style={{ flex: 1 }}>
-        <View style={{ padding: 15 }}>
-          <Text style={[styles.headerText, { marginBottom: 10, textAlign: 'center' }]}>
+      <ScrollView style={styles.scrollView}>
+        <View style={styles.contentContainer}>
+          <Text style={[styles.headerText, styles.centeredText]}>
             Help Center
           </Text>
-          <Text style={[styles.subHeaderText, { textAlign: 'center', marginBottom: 20 }]}>
+          <Text style={[styles.subHeaderText, styles.centeredText]}>
             Answers to your most common questions
           </Text>
 
           {helpSections.map((section, index) => (
             <View 
               key={index} 
-              style={{
-                backgroundColor: 'white',
-                borderRadius: 10,
-                marginBottom: 10,
-                shadowColor: '#000',
-                shadowOffset: { width: 0, height: 2 },
-                shadowOpacity: 0.1,
-                shadowRadius: 4,
-                elevation: 3
-              }}
+              style={styles.sectionContainer}
             >
               <TouchableOpacity 
                 onPress={() => toggleSection(index)}
-                style={{
-                  flexDirection: 'row',
-                  justifyContent: 'space-between',
-                  alignItems: 'center',
-                  padding: 15,
-                  borderBottomWidth: expandedSection === index ? 1 : 0,
-                  borderBottomColor: '#f0f0f0'
-                }}
+                style={[styles.sectionHeader, { borderBottomWidth: expandedSection === index ? 1 : 0 }]}
               >
                 <Text style={styles.tipTitle}>{section.title}</Text>
-                <Text style={{ color: '#fc8019' }}>
+                <Text style={styles.expandButton}>
                   {expandedSection === index ? '−' : '+'}
                 </Text>
               </TouchableOpacity>
 
               {expandedSection === index && (
-                <View style={{ padding: 15 }}>
+                <View style={styles.sectionContent}>
                   <Text style={styles.tipDescription}>{section.content}</Text>
                 </View>
               )}
             </View>
           ))}
 
-          <View style={{
-            backgroundColor: 'white',
-            borderRadius: 10,
-            padding: 15,
-            alignItems: 'center',
-            marginTop: 15,
-            shadowColor: '#000',
-            shadowOffset: { width: 0, height: 2 },
-            shadowOpacity: 0.1,
-            shadowRadius: 4,
-            elevation: 3
-          }}>
+          <View style={styles.contactContainer}>
             <Text style={styles.tipTitle}>Need More Help?</Text>
-            <Text style={[styles.tipDescription, { textAlign: 'center', marginVertical: 10 }]}>
+            <Text style={[styles.tipDescription, styles.centeredText]}>
               Contact our support team for further assistance
             </Text>
             <TouchableOpacity 
               onPress={handleContactSupport}
-              style={{
-                backgroundColor: '#fc8019',
-                paddingVertical: 10,
-                paddingHorizontal: 20,
-                borderRadius: 8
-              }}
+              style={styles.contactButton}
             >
-              <Text style={{ color: 'white', fontWeight: 'bold' }}>Contact Support</Text>
+              <Text style={styles.contactButtonText}>Contact Support</Text>
             </TouchableOpacity>
           </View>
         </View>
