+++ app/(home)/[AddonScreen].js
@@ -4,6 +4,7 @@ import { useLocalSearchParams, useRouter } from "expo-router";
 import { useSelector, useDispatch } from "react-redux";
 import { addAddonToItem, removeAddonFromItem } from "../../redux/CartReducer";
 import Menu from "../../data/MenuData.json"; // Assuming MenuData is stored here
+import { styles } from '../../Styles/AddOnStyles';  // Import styles
 
 const AddonScreen = () => {
   const { itemId, name, addons: initialAddons } = useLocalSearchParams();
@@ -53,59 +54,33 @@ const AddonScreen = () => {
   const renderAddonItem = ({ item }) => (
     <TouchableOpacity
       style={[
-        {
-          backgroundColor: "#fff",
-          padding: 12,
-          borderRadius: 8,
-          marginVertical: 8,
-          borderWidth: 1,
-          borderColor: "#ddd",
-          flexDirection: "row",
-          justifyContent: "space-between",
-          alignItems: "center",
-        },
-        selectedAddons.some((addon) => addon.id === item.id) && {
-          backgroundColor: "#f0f8ff",
-          borderColor: "#4caf50",
-          borderWidth: 2,
-        },
+        styles.addonItem,
+        selectedAddons.some((addon) => addon.id === item.id) && styles.addonItemSelected,
       ]}
       onPress={() => handleAddonSelection(item)}
     >
-      <Text style={{ fontSize: 16, color: "#333" }}>{item.name}</Text>
-      <Text style={{ fontSize: 16, color: "#666" }}>Rs. {item.price}</Text>
+      <Text style={styles.addonText}>{item.name}</Text>
+      <Text style={styles.addonPrice}>Rs. {item.price}</Text>
     </TouchableOpacity>
   );
 
   return (
-    <View style={{ flex: 1, backgroundColor: "#f8f8f8", padding: 16 }}>
+    <View style={styles.container}>
       {selectedItem ? (
         <>
-          <Text style={{ fontSize: 24, fontWeight: "bold", color: "#333", marginBottom: 20, textAlign: "center" }}>
-            Addons for {name}
-          </Text>
+          <Text style={styles.header}>Addons for {name}</Text>
           <FlatList
             data={itemAddons}
             keyExtractor={(item) => item.id.toString()}
             renderItem={renderAddonItem}
-            ListEmptyComponent={<Text>No addons available</Text>}
+            ListEmptyComponent={<Text style={styles.emptyText}>No addons available</Text>}
           />
-          <TouchableOpacity
-            style={{
-              backgroundColor: "#4caf50",
-              paddingVertical: 12,
-              borderRadius: 8,
-              marginTop: 20,
-              justifyContent: "center",
-              alignItems: "center",
-            }}
-            onPress={handleDone}
-          >
-            <Text style={{ fontSize: 18, color: "#fff", fontWeight: "bold" }}>Done</Text>
+          <TouchableOpacity style={styles.doneButton} onPress={handleDone}>
+            <Text style={styles.doneButtonText}>Done</Text>
           </TouchableOpacity>
         </>
       ) : (
-        <Text style={{ fontSize: 18, color: "#888", textAlign: "center" }}>Loading...</Text>
+        <Text style={styles.loadingText}>Loading...</Text>
       )}
     </View>
   );
