+++ package.json
@@ -17,7 +17,7 @@
     "@fortawesome/fontawesome-svg-core": "^6.6.0",
     "@fortawesome/react-native-fontawesome": "^0.3.2",
     "@react-native-async-storage/async-storage": "^1.23.1",
-    "@react-native-masked-view/masked-view": "^0.3.2",
+    "@react-native-masked-view/masked-view": "0.3.2",
     "@react-navigation/native": "^7.0.0",
     "@reduxjs/toolkit": "^2.3.0",
     "@rneui/themed": "^4.0.0-rc.8",
